Add support for GLES2 (#4)
This makes the game work with GLES2, which is way less buggy when exporting to HTML5 (aside from being buggy on the Web, apparently the GLES3 backend is now unmaintained in favor of the new Vulkan support that's coming in Godot 4).

For some reason the main level hard-crashed godot with a weird `Index p_index = 0 is out of bounds (size() = 0).At: ./core/cowdata.h:152` error, and it took me some tinkering to figure out that apparently what was causing it was the cat objective, and in particular the `Viewport` object in it, and something about the way it was configured.

I also figured out that there's an "Editable Children" checkbox next to imported scenes in a tree that makes all their children editable, so I moved the Cat Objective children to its scene.

The one major bummer here is that we're losing the GI probes, since those are only supported on the GLES3 backend.  And I tried adding baked lightmaps but it didn't seem to work--apparently using them is harder than I thought, so for now we're just increasing the amount of ambient light in the scene.