Add lamps (#6)
This adds some basic cube lamps on the floor in various areas to make the lighting more interesting.  Correspondingly, it vastly reduces the level of ambient light (it used to be very low prior to #4 because our GI Probes did a nice job of spreading indirect light across the scene, but then I had to increase it because GI Probes aren't supported on GLES2).

I also decided to disable shadows, because there was a huge amount of bias artifacts everywhere.  The attenuation of the lights is non-linear, which I think helped make the light look more natural and diffuse.

The challenging thing about disabling shadows is that it's easy for lighting to "bleed" into other rooms.  I've compensated for the obvious cases of this by simply placing lights on both sides of a wall when needed.

Note that this was originally based on #5, and I'm leaving the baked lightmap in the scene--just without any actual baking--so it's easy to tinker with in the future, in case I want to bring it back.